requirements.txt
discord==1.7.3
pytube==15.0.0
python-dotenv==0.19.1
ffmpeg-python==0.2.0
PyNaCl==1.5.0
